Explicit cast to int to avoid KeyErrors #22 (#23)
If null values are sampled in one time step (e.g. nodata classes or masked areas), the resulting column will be cast to float before the nans can be dropped. This can result in a KeyError when max classes are filtered. Explicitly casting to int after dropping nans solves the issue.

+++ sankee/sampling.py
@@ -43,7 +43,7 @@ def collect_sankey_data(
         else:
             raise e
 
-    data = pd.DataFrame.from_dict(features).dropna()
+    data = pd.DataFrame.from_dict(features).dropna().astype(int)
 
     for image in image_labels:
         if image not in data.columns:
